fix: 23 - fix error message not showing in registration inputs
