+++ index.html
@@ -8,5 +8,9 @@
   </head>
   <body>
     <script type="module" src="/src/main.ts"></script>
+    <!-- <h1>Hello</h1> -->
+    <h1>Hello-billing</h1>
+    <h1>Hello-shipping</h1>
+
   </body>
 </html>
