+++ src/pages/registration/constants.ts
@@ -1,8 +1,5 @@
 import type { Options } from '../../utils/create-elements/types';
 
-import { INPUT } from '../../shared/styles';
-
-export type RegistrationInputsProps = Omit<Options<'input'>, 'children' | 'parent' | 'tag' | 'text'>;
 type CountryOptions = Record<'base' | 'canada' | 'usa', CountryOptionsProps>;
 
 type CountryOptionsProps = Pick<Options<'option'>, 'attributes' | 'text'>;
@@ -12,10 +9,6 @@ type RegistrationCheckboxes = Record<
 >;
 
 type RegistrationCheckboxesProps = Omit<Options<'input'>, 'children' | 'parent' | 'tag' | 'text'>;
-type RegistrationInputs = Record<
-  'birthdate' | 'city' | 'firstname' | 'lastname' | 'postalcode' | 'street',
-  RegistrationInputsProps
->;
 
 export const REGISTRATION = {
   addressBlock: ['flex', 'flex-col', 'gap-1', 'border', 'border-[#252525]/50', 'p-2'],
@@ -32,99 +25,6 @@ export const REGISTRATION = {
     'border',
     'border-[#252525]/50',
   ],
-};
-
-export const REGISTRATION_INPUTS_CONFIG: RegistrationInputs = {
-  birthdate: {
-    attributes: {
-      name: 'date',
-      type: 'date',
-    },
-    classes: INPUT.registration,
-    events: {
-      change: (event) => {
-        if (event.target instanceof HTMLInputElement) {
-          console.log(event.target.value);
-        }
-      },
-    },
-  },
-  city: {
-    attributes: {
-      name: 'city',
-      placeholder: 'Enter city*',
-      type: 'text',
-    },
-    classes: INPUT.registration,
-    events: {
-      input: (event) => {
-        if (event.target instanceof HTMLInputElement) {
-          console.log(`city: ${event.target.value}`);
-        }
-      },
-    },
-  },
-  firstname: {
-    attributes: {
-      autocomplete: 'true',
-      name: 'firstname',
-      placeholder: 'Enter first name*',
-      type: 'text',
-    },
-    classes: INPUT.registration,
-    events: {
-      input: (event) => {
-        if (event.target instanceof HTMLInputElement) {
-          console.log(`first name: ${event.target.value}`);
-        }
-      },
-    },
-  },
-  lastname: {
-    attributes: {
-      name: 'lastname',
-      placeholder: 'Enter last name*',
-      type: 'text',
-    },
-    classes: INPUT.registration,
-    events: {
-      input: (event) => {
-        if (event.target instanceof HTMLInputElement) {
-          console.log(`last name: ${event.target.value}`);
-        }
-      },
-    },
-  },
-  postalcode: {
-    attributes: {
-      name: 'postalcode',
-      placeholder: 'Enter postal code*',
-      type: 'text',
-    },
-    classes: INPUT.registration,
-    events: {
-      input: (event) => {
-        if (event.target instanceof HTMLInputElement) {
-          console.log(`postal code: ${event.target.value}`);
-        }
-      },
-    },
-  },
-  street: {
-    attributes: {
-      name: 'street',
-      placeholder: 'Enter street*',
-      type: 'text',
-    },
-    classes: INPUT.registration,
-    events: {
-      input: (event) => {
-        if (event.target instanceof HTMLInputElement) {
-          console.log(`street: ${event.target.value}`);
-        }
-      },
-    },
-  },
 };
 
 export const countryOptions: CountryOptions = {
