+++ src/shared/ui-config/credential-inputs.ts
@@ -6,6 +6,8 @@ import { INPUT } from '../styles';
 export type BaseInputsProps = Omit<Options<'input'>, 'children' | 'parent' | 'tag' | 'text'>;
 type CredentialInputs = Record<'email' | 'password', BaseInputsProps>;
 
+import { createErrorMessage } from '../components/error-message';
+
 export const CREDENTIALS_INPUT_CONFIG: CredentialInputs = {
   email: {
     attributes: {
@@ -19,6 +21,7 @@ export const CREDENTIALS_INPUT_CONFIG: CredentialInputs = {
       input: (event) => {
         if (event.target instanceof HTMLInputElement) {
           console.log(`email: ${event.target.value}`);
+          createErrorMessage('hello', emailInput.errorContainer);
         }
       },
     },
